// template/react-starter-kit/src/.generated/api/model/messageChannel.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Nest Starter Kit
 * NestJS + MikroORM Starter Kit API
 * OpenAPI spec version: 1.0.0
 */

export type MessageChannel = typeof MessageChannel[keyof typeof MessageChannel];


export const MessageChannel = {
  EMAIL: 'EMAIL',
  SLACK: 'SLACK',
  IN_APP: 'IN_APP',
  SMS: 'SMS',
  ALIMTALK: 'ALIMTALK',
} as const;